+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Version vNext  (xxxx-xx-xx)
 - Issue #370: Relative path from a full path  (Thx QbProg)
 - Issue #373: Improve method `Directory.CreateDirectory` to allow creating a folder consisting only of spaces.
 - Issue #385: Correct applied fileSystemRights operator in method `File.Create()`.
+- Issue #408: Add `FileSystemEntryInfo.Extension` string property.
 
 ### Improvements